requirements.txt
streamlit
gitpython
flake8
coverage
pandas
pytest
pylint